Move ChangedPredicate to FluxBufferPredicate to avoid top level-class
