+++ reactor-core/src/main/java/reactor/core/publisher/Flux.java
@@ -2868,7 +2868,7 @@ public abstract class Flux<T> implements CorePublisher<T> {
 	 */
 	public final <V> Flux<List<T>> bufferUntilChanged(Function<? super T, ? extends V> keySelector,
 			BiPredicate<? super V, ? super V> keyComparator) {
-		Predicate<? super T> changed = new ChangedPredicate<T, V>(keySelector,
+		Predicate<? super T> changed = new FluxBufferPredicate.ChangedPredicate<T, V>(keySelector,
 				keyComparator);
 		return bufferUntil(changed, true);
 	}
